fix: move total field to common types
